Put literals in messages in curly quotes/monospace
This change causes curly quotation marks to be added around literals in the raw
error and warning messages (before the messages get passed to Velocity for
processing with the output-format-specific *.properties templates).

For non-"xhtml" output formats, the curly quotation marks are left as-is
(unprocessed) in the messages emitted to end users. But for the "xhtml" output
format, the curly quotation marks get changed into HTML `<code>` and `</code>`
tags, which causes their contents to get rendered in a monospace font.

Note: this change moves the HTML-escaping out of the xhtml.properties template
(`$esc.html`s added in f7766d87d) and back into the StyleSheetGenerator code
(making the existing `queryReplace()` method once again do the actual escaping).

This change also makes minor additions to the stylesheet for the CSS Validator
Web-based service, to cause other parts of the HTML output to get rendered in
monospace: the .unrecognized, .exp, .skippedString, and .codeContext classes.