Make parser handle “star property hack” gracefully
This change makes the parser gracefully handle cases like `*zoom: 1` where a CSS
identifier is prefixed with an asterisk (the so-called “star property hack”).
Without this change, the parser “derails” into a state where it ends up
incorrectly emitting spurious errors about later valid CSS constructs.

If the option to treat vendor extensions as warnings is set, a vendor-
extension warning message is reported; otherwise, a parse error is reported.

This change is itself a hack in that it achieves its effect by patching the
generated CssParser.java code after it’s been built by running javacc on the
CssParser.jj source.