Fix message bugs for selector after pseudo-element
This change fixes a couple bugs in the error message that’s reported for the
case of a selector that appears after a pseudo-element selector, and adds a
method to check for user-action pseudo classes

The first bug fixed by this change is, given a case like `p::first-line.foo`,
the CSS checker previously would report the following message:

> The pseudo-element .foo can't appear here in the context CSS level <level>

That message was wrong for that case because `.foo` isn’t a pseudo-element.

So this change causes the checker to instead report the following message:

> The selector .foo can't appear after the pseudo-element selector ::first-line

The second bug fixed by the change is, given a case like `p::first-line:hover`,
the checker would emit the following message:

> The pseudo-element :hover can't appear here in the context CSS level <level>

That message is wrong for two reasons: the first reason it’s wrong is the same
reason as above: `:hover` isn’t a pseudo-element — it's a pseudo-class; and the
second reason it’s wrong is that the current CSS Selectors spec allows it:

> A pseudo-element may be immediately followed by any combination of the user
action pseudo-classes

https://www.w3.org/TR/2013/WD-selectors4-20130502/#pseudo-elements

So, for handling such cases, this change adds a method for checking whether or
not particular pseudo-class is a user-action pseudo-class.

That method enables a check to determine that no error should be reported for
`p::first-line:hover` (because `hover` is a user-action pseudo-class), while
still correctly reporting an error for `p::first-line:visited` (because
`visited`) isn’t a user-action pseudo-class).